Synthetization

See merge request rootteam/rootcause!108

+++ core/src/main/java/fr/uge/modules/api/model/linking/Computation.java
@@ -1,4 +1,6 @@
 package fr.uge.modules.api.model.linking;
 
-public record Computation(String token_type, String value_log_first, String value_log_second, float proximity) {
+import java.math.BigDecimal;
+
+public record Computation(String token_type, String value_log_first, String value_log_second, BigDecimal proximity) {
 }
